[Kimmy] enable restartable

+++ dependency.sh
@@ -4,7 +4,7 @@ mkdir -p lib/
 
 cp rec-core/build/libs/rec-core-*.jar lib/
 
-wget https://github.com/kenpusney/dblite/releases/download/v0.0.1/dblite-0.0.1.jar -O lib/dblite.jar
-wget https://jdbc.postgresql.org/download/postgresql-42.1.1.jar -O lib/postgres-jdbc.jar
+[ -e lib/dblite.jar ] || wget https://github.com/kenpusney/dblite/releases/download/v0.0.1/dblite-0.0.1.jar -O lib/dblite.jar
+[ -e lib/postgres-jdbc.jar ] || wget https://jdbc.postgresql.org/download/postgresql-42.1.1.jar -O lib/postgres-jdbc.jar
